# Foil/Resources/AgentAccessOpenAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Foil Agent Access",
    "version": "1.2.0",
    "description": "Vocabulary discovery, in-memory preview, and inert proposals or action requests. Every change requires approval inside Foil; agents cannot approve changes."
  },
  "servers": [{"url": "http://foil", "description": "HTTP over Foil's owner-only Unix domain socket"}],
  "paths": {
    "/v1/instructions": {
      "get": {
        "operationId": "getInstructions",
        "responses": {
          "200": {"description": "Versioned model-readable instructions", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InstructionsResponse"}}}},
          "405": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/openapi.json": {
      "get": {
        "operationId": "getOpenAPI",
        "responses": {
          "200": {"description": "This OpenAPI document", "content": {"application/vnd.oai.openapi+json;version=3.1": {"schema": {"type": "object"}}}},
          "405": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/vocabulary/scopes": {
      "get": {
        "operationId": "listVocabularyScopes",
        "responses": {
          "200": {"description": "Allowed cleanup-scope identity and enabled state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScopesResponse"}}}},
          "405": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/vocabulary": {
      "get": {
        "operationId": "listVocabulary",
        "responses": {
          "200": {"description": "Allowed Vocabulary terms, corrections, and local-rule settings", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VocabularyResponse"}}}},
          "405": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/vocabulary/preview": {
      "post": {
        "operationId": "previewVocabularyCorrections",
        "description": "Validate and compile a hypothetical correction set in memory. Nothing is saved.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PreviewRequest"}}}},
        "responses": {
          "200": {"description": "Validation issues, normalized corrections, and synthetic examples", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PreviewResponse"}}}},
          "400": {"$ref": "#/components/responses/ErrorResponse"},
          "405": {"$ref": "#/components/responses/ErrorResponse"},
          "413": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/vocabulary/proposals": {
      "post": {
        "operationId": "proposeVocabularyCorrections",
        "description": "Validate and save an inert proposal for review in Foil. This does not apply Vocabulary changes.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProposalRequest"}}}},
        "responses": {
          "201": {"description": "Proposal created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProposalResponse"}}}},
          "200": {"description": "Idempotent replay of an existing proposal", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProposalResponse"}}}},
          "400": {"$ref": "#/components/responses/ErrorResponse"},
          "409": {"$ref": "#/components/responses/ErrorResponse"},
          "422": {"$ref": "#/components/responses/ErrorResponse"},
          "429": {"$ref": "#/components/responses/ErrorResponse"},
          "503": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/vocabulary/proposals/{proposal_id}": {
      "get": {
        "operationId": "getVocabularyProposalStatus",
        "description": "Return the durable disposition of a proposal.",
        "parameters": [{"name": "proposal_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Proposal status", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProposalResponse"}}}},
          "404": {"$ref": "#/components/responses/ErrorResponse"},
          "405": {"$ref": "#/components/responses/ErrorResponse"},
          "503": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/vocabulary/actions": {
      "post": {
        "operationId": "requestVocabularyAction",
        "description": "Save an inert request. Foil shows the exact change and scope for user approval. This endpoint never applies a change.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActionRequest"}}}},
        "responses": {
          "201": {"description": "Action request queued for Foil review", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActionResponse"}}}},
          "200": {"description": "Idempotent replay of the same request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActionResponse"}}}},
          "409": {"$ref": "#/components/responses/ErrorResponse"},
          "422": {"$ref": "#/components/responses/ErrorResponse"},
          "429": {"$ref": "#/components/responses/ErrorResponse"},
          "503": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    },
    "/v1/vocabulary/actions/{action_id}": {
      "get": {
        "operationId": "getVocabularyActionStatus",
        "parameters": [{"name": "action_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Durable user decision", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActionResponse"}}}},
          "404": {"$ref": "#/components/responses/ErrorResponse"},
          "503": {"$ref": "#/components/responses/ErrorResponse"}
        }
      }
    }
  },
  "components": {
    "responses": {
      "ErrorResponse": {
        "description": "Stable Agent Access error envelope",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      }
    },
    "schemas": {
      "InstructionsResponse": {
        "type": "object",
        "required": ["schema_version", "request_id", "service", "api_version", "available_operations", "bootstrap_command", "openapi_path", "openapi_command", "unavailable_behavior", "privacy", "limits"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string"},
          "service": {"type": "string"},
          "api_version": {"type": "string"},
          "available_operations": {"type": "array", "items": {"type": "string"}},
          "bootstrap_command": {"type": "string"},
          "openapi_path": {"type": "string", "const": "/v1/openapi.json"},
          "openapi_command": {"type": "string"},
          "unavailable_behavior": {"type": "string"},
          "privacy": {"type": "array", "items": {"type": "string"}},
          "limits": {"$ref": "#/components/schemas/Limits"}
        }
      },
      "Limits": {
        "type": "object",
        "required": ["maximum_header_bytes", "maximum_body_bytes", "maximum_correction_pairs", "maximum_spoken_forms_per_pair", "maximum_phrase_scalars", "request_deadline_seconds"],
        "properties": {
          "maximum_header_bytes": {"type": "integer"},
          "maximum_body_bytes": {"type": "integer"},
          "maximum_correction_pairs": {"type": "integer"},
          "maximum_spoken_forms_per_pair": {"type": "integer"},
          "maximum_phrase_scalars": {"type": "integer"},
          "request_deadline_seconds": {"type": "integer"}
        }
      },
      "VocabularyScope": {
        "type": "object",
        "required": ["id", "name", "is_default", "is_enabled"],
        "properties": {
          "id": {"type": "string"},
          "name": {"type": "string"},
          "is_default": {"type": "boolean"},
          "is_enabled": {"type": "boolean"}
        }
      },
      "VocabularyTerm": {
        "type": "object",
        "required": ["id", "term"],
        "properties": {
          "id": {"type": "string"},
          "term": {"type": "string"},
          "note": {"type": ["string", "null"]}
        }
      },
      "LocalRule": {
        "type": "object",
        "required": ["enabled", "case_sensitive"],
        "properties": {
          "enabled": {"type": "boolean"},
          "case_sensitive": {"type": "boolean"},
          "scope_id": {"type": ["string", "null"]}
        }
      },
      "VocabularyCorrection": {
        "type": "object",
        "required": ["id", "written_as", "correct_version"],
        "properties": {
          "id": {"type": "string"},
          "written_as": {"type": "string"},
          "correct_version": {"type": "string"},
          "note": {"type": ["string", "null"]},
          "local_rule": {"oneOf": [{"$ref": "#/components/schemas/LocalRule"}, {"type": "null"}]}
        }
      },
      "ScopesResponse": {
        "type": "object",
        "required": ["schema_version", "request_id", "scopes"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string"},
          "scopes": {"type": "array", "items": {"$ref": "#/components/schemas/VocabularyScope"}}
        }
      },
      "VocabularyResponse": {
        "type": "object",
        "required": ["schema_version", "request_id", "local_corrections_enabled", "terms", "corrections"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string"},
          "local_corrections_enabled": {"type": "boolean"},
          "terms": {"type": "array", "items": {"$ref": "#/components/schemas/VocabularyTerm"}},
          "corrections": {"type": "array", "items": {"$ref": "#/components/schemas/VocabularyCorrection"}}
        }
      },
      "PreviewRequest": {
        "type": "object",
        "required": ["corrections"],
        "properties": {"corrections": {"type": "array", "minItems": 1, "maxItems": 50, "items": {"$ref": "#/components/schemas/PreviewCorrection"}}}
      },
      "PreviewCorrection": {
        "type": "object",
        "required": ["spoken_forms", "replacement"],
        "properties": {
          "spoken_forms": {"type": "array", "minItems": 1, "maxItems": 10, "items": {"type": "string", "maxLength": 256}},
          "replacement": {"type": "string", "maxLength": 256},
          "scope_id": {"type": ["string", "null"]},
          "case_sensitive": {"type": "boolean", "default": false}
        }
      },
      "PreviewIssue": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "correction_index": {"type": ["integer", "null"]}
        }
      },
      "PreviewExample": {
        "type": "object",
        "required": ["input", "output", "replacement_count"],
        "properties": {
          "input": {"type": "string"},
          "output": {"type": "string"},
          "replacement_count": {"type": "integer"}
        }
      },
      "PreviewResponse": {
        "type": "object",
        "required": ["schema_version", "request_id", "valid", "issues", "normalized_corrections", "examples"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string"},
          "valid": {"type": "boolean"},
          "issues": {"type": "array", "items": {"$ref": "#/components/schemas/PreviewIssue"}},
          "normalized_corrections": {"type": "array", "items": {"$ref": "#/components/schemas/PreviewCorrection"}},
          "examples": {"type": "array", "items": {"$ref": "#/components/schemas/PreviewExample"}}
        }
      },
      "ProposalScope": {
        "type": "object",
        "required": ["kind", "id"],
        "properties": {
          "kind": {"type": "string", "enum": ["global", "cleanup_group"]},
          "id": {"type": "string", "description": "Use global for global scope, or an enabled scope ID returned by listVocabularyScopes."}
        }
      },
      "ProposalCorrection": {
        "type": "object",
        "required": ["spoken_forms", "replacement"],
        "properties": {
          "spoken_forms": {"type": "array", "minItems": 1, "maxItems": 10, "items": {"type": "string", "maxLength": 256}},
          "replacement": {"type": "string", "maxLength": 256},
          "note": {"type": ["string", "null"], "maxLength": 256},
          "case_sensitive": {"type": "boolean", "default": false}
        }
      },
      "ProposalRequest": {
        "type": "object",
        "required": ["schema_version", "request_id", "scope", "corrections"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "scope": {"$ref": "#/components/schemas/ProposalScope"},
          "corrections": {"type": "array", "minItems": 1, "maxItems": 50, "items": {"$ref": "#/components/schemas/ProposalCorrection"}}
        }
      },
      "ProposalResponse": {
        "type": "object",
        "required": ["schema_version", "request_id", "client_request_id", "proposal_id", "state", "replayed", "created_at", "updated_at"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string"},
          "client_request_id": {"type": "string"},
          "proposal_id": {"type": "string", "format": "uuid"},
          "state": {"type": "string", "enum": ["pending", "applied", "rejected", "discarded"]},
          "replayed": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "ActionRequest": {
        "type": "object",
        "required": ["schema_version", "request_id", "action"],
        "description": "Specify only the fields for one action. apply_proposal: proposal_id; set_local_corrections_enabled: enabled; set_correction_scope: correction_id and scope_id (global or enabled group ID); assign_app_to_group: app_bundle_id and group_id (enabled group ID). App assignment also changes that app's Cleanup Group routing. No approval claim is accepted in this request.",
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string", "minLength": 1, "maxLength": 128},
          "action": {"type": "string", "enum": ["apply_proposal", "set_local_corrections_enabled", "set_correction_scope", "assign_app_to_group"]},
          "proposal_id": {"type": "string", "format": "uuid"},
          "enabled": {"type": "boolean"},
          "correction_id": {"type": "string", "format": "uuid"},
          "scope_id": {"type": "string"},
          "app_bundle_id": {"type": "string"},
          "group_id": {"type": "string"}
        }
      },
      "ActionResponse": {
        "type": "object",
        "required": ["schema_version", "request_id", "client_request_id", "action_id", "state", "replayed", "created_at", "updated_at"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string"},
          "client_request_id": {"type": "string"},
          "action_id": {"type": "string", "format": "uuid"},
          "state": {"type": "string", "enum": ["pending", "approved_pending_apply", "approved", "rejected", "cancelled_after_approval"]},
          "replayed": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"},
          "approved_at": {"type": ["string", "null"], "format": "date-time"}
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {"code": {"type": "string"}, "message": {"type": "string"}}
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["schema_version", "request_id", "error"],
        "properties": {
          "schema_version": {"type": "integer", "const": 1},
          "request_id": {"type": "string"},
          "error": {"$ref": "#/components/schemas/ErrorDetail"}
        }
      }
    }
  },
  "x-foil-limits": {
    "maximum_header_bytes": 16384,
    "maximum_body_bytes": 65536,
    "maximum_correction_pairs": 50,
    "maximum_spoken_forms_per_pair": 10,
    "maximum_phrase_scalars": 256,
    "request_deadline_seconds": 5,
    "bootstrap_max_time_seconds": 12
  },
  "x-foil-proposal-states": ["pending", "applied", "rejected", "discarded"],
  "x-foil-action-states": ["pending", "approved_pending_apply", "approved", "rejected", "cancelled_after_approval"],
  "x-foil-privacy": {
    "same_user_readable": ["scope_id", "scope_name", "scope_enabled", "term", "correction", "note", "local_rule_enabled", "case_sensitive"],
    "excludes": ["history", "transcript", "audio", "credentials", "provider_configuration", "source_record", "source_app", "timestamps", "project_files", "repository", "clipboard", "active_application"]
  }
}